refactor: rename PeopleCard to Pascal case
