+++ app/about/page.tsx
@@ -4,7 +4,7 @@ import { Hero } from "@/components/Hero"
 import { TextAnimate } from "@/components/magicui/text-animate"
 import { SectionHeader } from "@/components/ui/section-header"
 import { Card, CardContent } from "@/components/ui/Card"
-import { CardWithImage } from "@/components/ui/people-card"
+import { CardWithImage } from "@/components/ui/PeopleCard"
 import { readContentFile } from "@/lib/content-utils"
 import { Workday } from "@/components/Workday"
 import { getWorkdayData } from "@/app/about/queries"
