+++ components/ui/PeopleCard.tsx
